Refactor: move test package in right position
